consolidate duplicated functions
