+++ scripts/lib/CIME/XML/archive.py
@@ -3,13 +3,13 @@ Interface to the archive.xml file.  This class inherits from GenericXML.py
 """
 
 from CIME.XML.standard_module_setup import *
-from CIME.XML.generic_xml import GenericXML
+from CIME.XML.archive_base import ArchiveBase
 from CIME.XML.files import Files
 from CIME.utils import expect, get_model
 
 logger = logging.getLogger(__name__)
 
-class Archive(GenericXML):
+class Archive(ArchiveBase):
 
     def __init__(self, infile=None, files=None):
         """
@@ -18,8 +18,7 @@ class Archive(GenericXML):
         if files is None:
             files = Files()
         schema = files.get_schema("ARCHIVE_SPEC_FILE")
-
-        GenericXML.__init__(self, infile, schema)
+        super(Archive, self).__init__(infile, schema)
 
     def setup(self, env_archive, components, files=None):
         if files is None:
@@ -80,27 +79,3 @@ class Archive(GenericXML):
                     config_archive_files.append(compval)
 
         return list(set(config_archive_files))
-
-    def get_entry(self, compname):
-        return self.get_optional_child('comp_archive_spec',
-                                       attributes={"compname":compname})
-
-    def get_rest_file_extensions(self, archive_entry):
-        file_extensions = []
-        nodes = self.get_children('rest_file_extension', root=archive_entry)
-        for node in nodes:
-            file_extensions.append(self.text(node))
-        return file_extensions
-
-    def get_hist_file_extensions(self, archive_entry):
-        file_extensions = []
-        nodes = self.get_children('hist_file_extension', root=archive_entry)
-        for node in nodes:
-            file_extensions.append(self.text(node))
-        return file_extensions
-
-    def get_entry_value(self, name, archive_entry):
-        node = self.get_optional_child(name, root=archive_entry)
-        if node is not None:
-            return self.text(node)
-        return None
