+++ src/drivers/mct/cime_config/config_archive.xml
@@ -8,7 +8,10 @@
       <rpointer_content >$CASE.cpl$NINST_STRING.r.$DATENAME.nc</rpointer_content>
     </rpointer>
     <test_file_names>
+      <tfile disposition="move">cpl_0001.log.5548574.chadmin1.180228-124723.gz</tfile>
       <tfile disposition="copy">casename.cpl.r.1976-01-01-00000.nc</tfile>
+      <tfile disposition="copy">rpointer.drv_0001</tfile>
+      <tfile disposition="copy">rpointer.drv</tfile>
       <tfile disposition="ignore">casenamenot.cpl.r.1976-01-01-00000.nc</tfile>
     </test_file_names>
   </comp_archive_spec>
